+++ src/components/dapp-staking/my-staking/styles/my-rewards.scss
@@ -18,7 +18,7 @@
 }
 
 .wrapper--icon-help {
-  padding-top: 7px;
+  margin-top: -1px;
 }
 
 .card {
@@ -73,6 +73,13 @@
     line-height: 18px;
     color: $gray-5-selected;
   }
+  .value--claim {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    row-gap: 2px;
+    margin-right: 0px;
+  }
 }
 
 .explorer-icon {
@@ -84,6 +91,14 @@
   &:hover {
     stroke: white;
   }
+}
+
+.text--eras {
+  font-size: 14px;
+}
+
+.text--rewards-amount {
+  font-size: 18px;
 }
 
 .body--dark {
