+++ src/i18n/en-US/index.ts
@@ -193,7 +193,9 @@ export default {
     turnOff: 'Turn Off',
     totalEarned: 'Total Earned (all-time)',
     availableToClaimTip:
-      'Number of eras that is shown here is per dApp. The maximum number of eras you can claim at once is 50. You may need to claim multiple times if you leave it too long.',
+      "The amount of pending rewards(estimated) that is shown here is for Staker's. The Number of eras that is shown here is per dApp.",
+    availableToClaimTip2:
+      'The maximum number of eras you can claim at once is 50. You may need to claim multiple times if you leave it too long.',
     restakeTip:
       'By turning on, your rewards will be automatically re-staked when you make a claim.',
   },
