+++ src/modules/dapp-staking/index.ts
@@ -1,5 +1,7 @@
 export { checkIsLimitedProvider } from 'src/modules/dapp-staking/utils';
 
+export { getPendingRewards } from 'src/modules/dapp-staking/pending-rewards';
+
 export interface StakingData {
   address: string;
   name: string;
